backend/requirements.txt
flask==3.0.3
torch==2.3.0
torchvision==0.18.0
Pillow==10.4.0
numpy